Split SPI stuff to a separate file
