+++ components/board/genericlinux/board.go
@@ -10,7 +10,6 @@ import (
 	"fmt"
 	"strconv"
 	"sync"
-	"sync/atomic"
 	"time"
 
 	"github.com/edaniels/golog"
@@ -22,8 +21,6 @@ import (
 	"periph.io/x/conn/v3/gpio"
 	"periph.io/x/conn/v3/gpio/gpioreg"
 	"periph.io/x/conn/v3/physic"
-	"periph.io/x/conn/v3/spi"
-	"periph.io/x/conn/v3/spi/spireg"
 
 	"go.viam.com/rdk/components/board"
 	"go.viam.com/rdk/grpc"
@@ -387,62 +384,6 @@ type pwmSetting struct {
 func (b *sysfsBoard) SPIByName(name string) (board.SPI, bool) {
 	s, ok := b.spis[name]
 	return s, ok
-}
-
-type spiBus struct {
-	mu         sync.Mutex
-	openHandle *spiHandle
-	bus        atomic.Pointer[string]
-}
-
-type spiHandle struct {
-	bus      *spiBus
-	isClosed bool
-}
-
-func (sb *spiBus) OpenHandle() (board.SPIHandle, error) {
-	sb.mu.Lock()
-	sb.openHandle = &spiHandle{bus: sb, isClosed: false}
-	return sb.openHandle, nil
-}
-
-func (sb *spiBus) Close(ctx context.Context) error {
-	return nil
-}
-
-func (sb *spiBus) reset(bus string) {
-	sb.bus.Store(&bus)
-}
-
-func (sh *spiHandle) Xfer(ctx context.Context, baud uint, chipSelect string, mode uint, tx []byte) (rx []byte, err error) {
-	if sh.isClosed {
-		return nil, errors.New("can't use Xfer() on an already closed SPIHandle")
-	}
-
-	busPtr := sh.bus.bus.Load()
-	if busPtr == nil {
-		return nil, errors.New("no bus selected")
-	}
-
-	port, err := spireg.Open(fmt.Sprintf("SPI%s.%s", *busPtr, chipSelect))
-	if err != nil {
-		return nil, err
-	}
-	defer func() {
-		err = multierr.Combine(err, port.Close())
-	}()
-	conn, err := port.Connect(physic.Hertz*physic.Frequency(baud), spi.Mode(mode), 8)
-	if err != nil {
-		return nil, err
-	}
-	rx = make([]byte, len(tx))
-	return rx, conn.Tx(tx, rx)
-}
-
-func (sh *spiHandle) Close() error {
-	sh.isClosed = true
-	sh.bus.mu.Unlock()
-	return nil
 }
 
 func (b *sysfsBoard) I2CByName(name string) (board.I2C, bool) {
